go: move handlers to civssession pkg
